+++ client/transaction-pool/src/tests/mod.rs
@@ -14,9 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with Substrate.  If not, see <http://www.gnu.org/licenses/>.
 
-use super::*;
+mod api;
 
-use crate::{BasicPool, MaintainedTransactionPool};
+use crate::*;
 use codec::Encode;
 use futures::executor::block_on;
 use parking_lot::RwLock;
@@ -31,124 +31,7 @@ use substrate_test_runtime_client::{
 	runtime::{AccountId, Block, BlockNumber, Extrinsic, Hash, Header, Index, Transfer},
 	AccountKeyring::{self, *},
 };
-
-struct TestApi {
-	pub modifier: RwLock<Box<dyn Fn(&mut ValidTransaction) + Send + Sync>>,
-	pub chain_block_by_number: RwLock<HashMap<BlockNumber, Vec<Extrinsic>>>,
-	pub chain_headers_by_number: RwLock<HashMap<BlockNumber, Header>>,
-	pub invalid_hashes: RwLock<HashSet<ExHash<Self>>>,
-	pub validation_requests: RwLock<Vec<Extrinsic>>,
-}
-
-impl TestApi {
-	fn default() -> Self {
-		TestApi {
-			modifier: RwLock::new(Box::new(|_| {})),
-			chain_block_by_number: RwLock::new(HashMap::new()),
-			invalid_hashes: RwLock::new(HashSet::new()),
-			chain_headers_by_number: RwLock::new(HashMap::new()),
-			validation_requests: RwLock::new(Default::default()),
-		}
-	}
-
-	fn push_block(&self, block_number: BlockNumber, xts: Vec<Extrinsic>) {
-		self.chain_block_by_number.write().insert(block_number, xts);
-		self.chain_headers_by_number.write().insert(block_number, Header {
-			number: block_number,
-			digest: Default::default(),
-			extrinsics_root:  Default::default(),
-			parent_hash: Default::default(),
-			state_root: Default::default(),
-		});
-	}
-}
-
-impl sc_transaction_graph::ChainApi for TestApi {
-	type Block = Block;
-	type Hash = Hash;
-	type Error = error::Error;
-	type ValidationFuture = futures::future::Ready<error::Result<TransactionValidity>>;
-	type BodyFuture = futures::future::Ready<error::Result<Option<Vec<Extrinsic>>>>;
-
-	fn validate_transaction(
-		&self,
-		at: &BlockId<Self::Block>,
-		uxt: sc_transaction_graph::ExtrinsicFor<Self>,
-	) -> Self::ValidationFuture {
-
-		self.validation_requests.write().push(uxt.clone());
-
-		let expected = index(at);
-		let requires = if expected == uxt.transfer().nonce {
-			vec![]
-		} else {
-			vec![vec![uxt.transfer().nonce as u8 - 1]]
-		};
-		let provides = vec![vec![uxt.transfer().nonce as u8]];
-
-		if self.invalid_hashes.read().contains(&self.hash_and_length(&uxt).0) {
-			return futures::future::ready(Ok(
-				Err(TransactionValidityError::Invalid(InvalidTransaction::Custom(0)))
-			))
-		}
-
-		let mut validity = ValidTransaction {
-			priority: 1,
-			requires,
-			provides,
-			longevity: 64,
-			propagate: true,
-		};
-
-		(self.modifier.read())(&mut validity);
-
-		futures::future::ready(Ok(Ok(validity)))
-	}
-
-	fn block_id_to_number(
-		&self,
-		at: &BlockId<Self::Block>,
-	) -> error::Result<Option<sc_transaction_graph::NumberFor<Self>>> {
-		Ok(Some(number_of(at)))
-	}
-
-	fn block_id_to_hash(
-		&self,
-		at: &BlockId<Self::Block>,
-	) -> error::Result<Option<sc_transaction_graph::BlockHash<Self>>> {
-		Ok(match at {
-			generic::BlockId::Hash(x) => Some(x.clone()),
-			_ => Some(Default::default()),
-		})
-	}
-
-	fn hash_and_length(
-		&self,
-		ex: &sc_transaction_graph::ExtrinsicFor<Self>,
-	) -> (Self::Hash, usize) {
-		let encoded = ex.encode();
-		(BlakeTwo256::hash(&encoded), encoded.len())
-	}
-
-	fn block_body(&self, id: &BlockId<Self::Block>) -> Self::BodyFuture {
-		futures::future::ready(Ok(if let BlockId::Number(num) = id {
-			self.chain_block_by_number.read().get(num).cloned()
-		} else {
-			None
-		}))
-	}
-}
-
-fn index(at: &BlockId<Block>) -> u64 {
-	209 + number_of(at)
-}
-
-fn number_of(at: &BlockId<Block>) -> u64 {
-	match at {
-		generic::BlockId::Number(n) => *n as u64,
-		_ => 0,
-	}
-}
+use api::TestApi;
 
 fn uxt(who: AccountKeyring, nonce: Index) -> Extrinsic {
 	let transfer = Transfer {
@@ -172,7 +55,7 @@ fn maintained_pool() -> BasicPool<TestApi, Block> {
 #[test]
 fn submission_should_work() {
 	let pool = pool();
-	assert_eq!(209, index(&BlockId::number(0)));
+	assert_eq!(209, api::index(&BlockId::number(0)));
 	block_on(pool.submit_one(&BlockId::number(0), uxt(Alice, 209))).unwrap();
 
 	let pending: Vec<_> = pool.ready().map(|a| a.data.transfer().nonce).collect();
@@ -245,7 +128,7 @@ fn should_ban_invalid_transactions() {
 #[test]
 fn should_correctly_prune_transactions_providing_more_than_one_tag() {
 	let api = TestApi::default();
-	*api.modifier.write() = Box::new(|v: &mut ValidTransaction| {
+	*api.valid_modifier.write() = Box::new(|v: &mut ValidTransaction| {
 		v.provides.push(vec![155]);
 	});
 	let pool = Pool::new(Default::default(), Arc::new(api));
@@ -288,7 +171,6 @@ fn should_prune_old_during_maintenance() {
 	assert_eq!(pool.status().ready, 0);
 }
 
-
 #[test]
 fn should_revalidate_during_maintenance() {
 	let xt1 = uxt(Alice, 209);
@@ -298,12 +180,12 @@ fn should_revalidate_during_maintenance() {
 	block_on(pool.submit_one(&BlockId::number(0), xt1.clone())).expect("1. Imported");
 	block_on(pool.submit_one(&BlockId::number(0), xt2.clone())).expect("2. Imported");
 	assert_eq!(pool.status().ready, 2);
-	assert_eq!(pool.api.validation_requests.read().len(), 2);
+	assert_eq!(pool.api.validation_requests().len(), 2);
 
 	pool.api.push_block(1, vec![xt1.clone()]);
 
 	block_on(pool.maintain(&BlockId::number(1), &[]));
 	assert_eq!(pool.status().ready, 1);
 	// test that pool revalidated transaction that left ready and not included in the block
-	assert_eq!(pool.api.validation_requests.read().len(), 3);
+	assert_eq!(pool.api.validation_requests().len(), 3);
 }
